initial tests passing, added coverage, still needs work
