+++ .travis.yml
@@ -36,5 +36,4 @@ script:
   - npm run lint
   - npm run test
   - npm run build
-  - npm run test-e2e
   - npm run package
