+++ appveyor.yml
@@ -1,6 +1,6 @@
 os: unstable
 cache:
-  - node_modules 
+  - node_modules
 environment:
   matrix:
     - nodejs_version: 6
@@ -24,5 +24,4 @@ test_script:
   - npm run lint
   - npm run test
   - npm run build
-  - npm run test-e2e
   - npm run package
