--- test/containers/CounterPage.spec.js
@@ -1,55 +1,0 @@
-import { expect } from 'chai';
-import React from 'react';
-import { mount } from 'enzyme';
-import { Provider } from 'react-redux';
-import CounterPage from '../../app/containers/CounterPage';
-import configureStore from '../../app/store/configureStore';
-
-
-function setup(initialState) {
-  const store = configureStore(initialState);
-  const app = mount(
-    <Provider store={store}>
-      <CounterPage />
-    </Provider>
-  );
-  return {
-    app,
-    buttons: app.find('button'),
-    p: app.find('.counter')
-  };
-}
-
-
-describe('containers', () => {
-  describe('App', () => {
-    it('should display initial count', () => {
-      const { p } = setup();
-      expect(p.text()).to.match(/^0$/);
-    });
-
-    it('should display updated count after increment button click', () => {
-      const { buttons, p } = setup();
-      buttons.at(0).simulate('click');
-      expect(p.text()).to.match(/^1$/);
-    });
-
-    it('should display updated count after descrement button click', () => {
-      const { buttons, p } = setup();
-      buttons.at(1).simulate('click');
-      expect(p.text()).to.match(/^-1$/);
-    });
-
-    it('shouldnt change if even and if odd button clicked', () => {
-      const { buttons, p } = setup();
-      buttons.at(2).simulate('click');
-      expect(p.text()).to.match(/^0$/);
-    });
-
-    it('should change if odd and if odd button clicked', () => {
-      const { buttons, p } = setup({ counter: 1 });
-      buttons.at(2).simulate('click');
-      expect(p.text()).to.match(/^2$/);
-    });
-  });
-});
